fix import/no-cycle lint
