+++ .eslintrc.js
@@ -1,3 +1,7 @@
+const typeScriptExtensions = [".ts", ".cts", ".mts", ".tsx"];
+
+const allExtensions = [...typeScriptExtensions, ".js", ".jsx"];
+
 module.exports = {
   root: true,
   extends: [
@@ -8,6 +12,18 @@ module.exports = {
   ],
   parser: "@typescript-eslint/parser",
   plugins: ["import", "@typescript-eslint"],
+  settings: {
+    "import/extensions": allExtensions,
+    "import/external-module-folders": ["node_modules", "node_modules/@types"],
+    "import/parsers": {
+      "@typescript-eslint/parser": typeScriptExtensions,
+    },
+    "import/resolver": {
+      node: {
+        extensions: allExtensions,
+      },
+    },
+  },
   rules: {
     curly: 1,
     "import/no-extraneous-dependencies": [
@@ -23,17 +39,19 @@ module.exports = {
     "@typescript-eslint/no-non-null-assertion": "off",
     "@typescript-eslint/no-explicit-any": "off",
     "@typescript-eslint/ban-ts-comment": "off",
-    "import/no-restricted-paths": [
-      "error",
-      {
-        zones: [
-          {
-            target: "types/**/*",
-            from: "src/**/*",
-            message: "Import from this module to types is not allowed.",
-          },
-        ],
-      },
-    ],
+    "import/no-cycle": "error",
+    // doesn't work:
+    // "import/no-restricted-paths": [
+    //   "error",
+    //   {
+    //     zones: [
+    //       {
+    //         target: "./src/**/*",
+    //         from: "./types/**/*",
+    //         message: "Import from this module to types is not allowed.",
+    //       },
+    //     ],
+    //   },
+    // ],
   },
 };
