+++ packages/core/src/api/exporters/html/externalHTMLExporter.ts
@@ -3,7 +3,7 @@ import rehypeParse from "rehype-parse";
 import rehypeStringify from "rehype-stringify";
 import { unified } from "unified";
 
-import { BlockNoteEditor } from "../../../editor/BlockNoteEditor";
+import type { BlockNoteEditor } from "../../../editor/BlockNoteEditor";
 import { BlockSchema, PartialBlock } from "../../../schema/blocks/types";
 import { InlineContentSchema } from "../../../schema/inlineContent/types";
 import { StyleSchema } from "../../../schema/styles/types";
