+++ packages/core/src/api/exporters/html/internalHTMLSerializer.ts
@@ -1,5 +1,5 @@
 import { DOMSerializer, Fragment, Node, Schema } from "prosemirror-model";
-import { BlockNoteEditor } from "../../../editor/BlockNoteEditor";
+import type { BlockNoteEditor } from "../../../editor/BlockNoteEditor";
 import { BlockSchema, PartialBlock } from "../../../schema/blocks/types";
 import { InlineContentSchema } from "../../../schema/inlineContent/types";
 import { StyleSchema } from "../../../schema/styles/types";
