+++ packages/core/src/api/exporters/html/util/sharedHTMLConversion.ts
@@ -1,6 +1,6 @@
 import { DOMSerializer, Fragment, Node } from "prosemirror-model";
 
-import { BlockNoteEditor } from "../../../../editor/BlockNoteEditor";
+import type { BlockNoteEditor } from "../../../../editor/BlockNoteEditor";
 import { BlockSchema } from "../../../../schema/blocks/types";
 import { InlineContentSchema } from "../../../../schema/inlineContent/types";
 import { StyleSchema } from "../../../../schema/styles/types";
