+++ packages/core/src/extensions/SideMenu/SideMenuPlugin.ts
@@ -2,16 +2,16 @@ import { PluginView } from "@tiptap/pm/state";
 import { Node } from "prosemirror-model";
 import { NodeSelection, Plugin, PluginKey, Selection } from "prosemirror-state";
 import { EditorView } from "prosemirror-view";
-import { BlockNoteEditor } from "../../editor/BlockNoteEditor";
+import { EventEmitter } from "../../@util/EventEmitter";
 import { createExternalHTMLExporter } from "../../api/exporters/html/externalHTMLExporter";
 import { createInternalHTMLSerializer } from "../../api/exporters/html/internalHTMLSerializer";
 import { cleanHTMLToMarkdown } from "../../api/exporters/markdown/markdownExporter";
+import { getBlockInfoFromPos } from "../../api/getBlockInfoFromPos";
+import type { BlockNoteEditor } from "../../editor/BlockNoteEditor";
 import { Block, BlockSchema } from "../../schema/blocks/types";
 import { InlineContentSchema } from "../../schema/inlineContent/types";
 import { StyleSchema } from "../../schema/styles/types";
 import { BaseUiElementState } from "../@shared/BaseUiElementTypes";
-import { EventEmitter } from "../../@util/EventEmitter";
-import { getBlockInfoFromPos } from "../../api/getBlockInfoFromPos";
 import { slashMenuPluginKey } from "../SlashMenu/SlashMenuPlugin";
 import { MultipleNodeSelection } from "./MultipleNodeSelection";
