+++ packages/react/src/components/SideMenu/DefaultButtons/AddBlockButton.tsx
@@ -1,7 +1,7 @@
 import { BlockSchema } from "@blocknote/core";
 import { AiOutlinePlus } from "react-icons/ai";
 import { SideMenuButton } from "../SideMenuButton";
-import { SideMenuProps } from "../SideMenuPositioner";
+import type { SideMenuProps } from "../SideMenuPositioner";
 
 export const AddBlockButton = <BSchema extends BlockSchema>(
   props: SideMenuProps<BSchema, any, any>
