+++ packages/react/src/components/SideMenu/DefaultButtons/DragHandle.tsx
@@ -3,7 +3,7 @@ import { Menu } from "@mantine/core";
 import { MdDragIndicator } from "react-icons/md";
 import { DefaultDragHandleMenu } from "../DragHandleMenu/DefaultDragHandleMenu";
 import { SideMenuButton } from "../SideMenuButton";
-import { SideMenuProps } from "../SideMenuPositioner";
+import type { SideMenuProps } from "../SideMenuPositioner";
 
 export const DragHandle = <BSchema extends BlockSchema>(
   props: SideMenuProps<BSchema, any, any>
