+++ packages/react/src/components/SideMenu/DefaultSideMenu.tsx
@@ -4,7 +4,7 @@ import { StyleSchema } from "@blocknote/core";
 import { AddBlockButton } from "./DefaultButtons/AddBlockButton";
 import { DragHandle } from "./DefaultButtons/DragHandle";
 import { SideMenu } from "./SideMenu";
-import { SideMenuProps } from "./SideMenuPositioner";
+import type { SideMenuProps } from "./SideMenuPositioner";
 
 export const DefaultSideMenu = <
   BSchema extends BlockSchema,
